// src/app/models/productPlaceholder.model.ts
export interface ProductPlaceholder {
    id: string;
    name: string;
    description: string;
    metacritic_score: number; 
    rating: number;
    released: string;
    background_image: string;
    background_image_additional: string;
    website: string;
    rawg_id: number;
    genres: string; 
    platforms: string;
    developers: string;
    publishers: string;
    screenshots: string[]; 
    esrb_rating: string; 
    playtime: number;
    achievements_count: number;
    parent_platforms: string;

} 